+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 - Bring most hard-coded constants to top of file
 - Enable Python 3.14 in GH CI
 - Investigate tests on GH failing due to different linear solver
+- Disallow conversions and surplus in last 2 years
 
 ### Version 2026.01.17
 - Add more parameters to control solver